# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BacklogRow": {
        "properties": {
          "accepting": {
            "title": "Accepting",
            "type": "boolean"
          },
          "grade": {
            "title": "Grade",
            "type": "integer"
          },
          "max_open_cases": {
            "title": "Max Open Cases",
            "type": "integer"
          },
          "open_cases": {
            "title": "Open Cases",
            "type": "integer"
          },
          "open_permits": {
            "description": "Started, unsubmitted written tests that may still become cases",
            "title": "Open Permits",
            "type": "integer"
          },
          "overdue_cases": {
            "description": "Queued cases past their due time (service obligation missed)",
            "title": "Overdue Cases",
            "type": "integer"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "grade",
          "subject",
          "max_open_cases",
          "open_cases",
          "open_permits",
          "accepting",
          "overdue_cases"
        ],
        "title": "BacklogRow",
        "type": "object"
      },
...
      "CapacityIn": {
        "properties": {
          "max_open_cases": {
            "description": "0 stops new written starts in this scope",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Max Open Cases",
            "type": "integer"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 5,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "max_open_cases",
          "reason"
        ],
        "title": "CapacityIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Authoritative backend for student web, Android, iOS and staff admin.",
    "title": "Punjab Learning Portal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ops/written-capacity/{grade}/{subject}": {
      "put": {
        "operationId": "set_capacity_v1_ops_written_capacity__grade___subject__put",
        "parameters": [
          {
            "in": "path",
            "name": "grade",
            "required": true,
            "schema": {
              "title": "Grade",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "subject",
            "required": true,
            "schema": {
              "title": "Subject",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapacityIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacklogRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set funded teacher-review capacity for a class and subject (operators, MFA, audited)",
        "tags": [
          "written marking"
        ]
      }
    },
...
    "/v1/studio/items/{item_id}/review": {
      "post": {
        "operationId": "review_v1_studio_items__item_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review",
        "tags": [
          "studio"
        ]
      }
    },
...
    "/v1/studio/written/cases/{case_id}": {
      "get": {
        "operationId": "case_detail_v1_studio_written_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Case Detail",
        "tags": [
          "written marking"
        ]
      }
    },
    "/v1/studio/written/cases/{case_id}/decision": {
      "post": {
        "operationId": "decision_v1_studio_written_cases__case_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save or release marks (expected-version CAS; awards must be permitted rubric levels)",
        "tags": [
          "written marking"
        ]
      }
    },
    "/v1/studio/written/cases/{case_id}/lease": {
      "post": {
        "operationId": "take_lease_v1_studio_written_cases__case_id__lease_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Take or renew the marking lease",
        "tags": [
          "written marking"
        ]
      }
    },
    "/v1/studio/written/cases/{case_id}/pages/{page_id}": {
      "get": {
        "operationId": "case_page_v1_studio_written_cases__case_id__pages__page_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submitted evidence page (markers in scope)",
        "tags": [
          "written marking"
        ]
      }
    },
...
    "/v1/studio/written/cases/{case_id}/rebase": {
      "post": {
        "operationId": "rebase_v1_studio_written_cases__case_id__rebase_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebaseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebase an open recheck onto the current result (academic adjudicators; audited)",
        "tags": [
          "written marking"
        ]
      }
    },
...
  }
}